feat: rename create cluster to apply
