+++ cmd/download.go
@@ -11,7 +11,7 @@ import (
 	"github.com/sighupio/furyctl/internal/analytics"
 )
 
-func NewDownloadCmd(tracker *analytics.Tracker) *cobra.Command {
+func NewDownloadCommand(tracker *analytics.Tracker) *cobra.Command {
 	dumpCmd := &cobra.Command{
 		Use:   "download",
 		Short: "Download all dependencies for the Kubernetes Fury Distribution version specified in the configuration file",
